WARNING huge "template" deviation from the course
In lesson 50, Jon talks about 2 ways of working with layouts.
The first is including the "header" and "footer" templates in every page. This is the method Jon said he would use for the rest of the course.
The second is having a central layout and passing different "page" templates.

I am going to tread dangerous waters and try to proceed using the 2nd method.